requirements.txt
streamlit
tensorflow
scikit-learn
pandas
numpy
matplotlib
seaborn
joblib
mysql-connector-python